pyproject.toml
[tool.poetry]
name = "anonymization_verification"
version = "0.1.0"
description = "A beautiful description."
authors = ["Reinder Vos de Wael <reinder.vosdewael@childmind.org>"]
license = "LGPL-2.1"
readme = "README.md"
packages = [{include = "anonymization_verification", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-docx = "^0.8.11"
spacy = "^3.6.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
mypy = "^1.4.0"
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"
pytest-cov = "^4.1.0"

[tool.poetry.group.docs.dependencies]
pdoc = "^14.0.0"

[tool.poetry.extras]
docs = ["pdoc"]

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]

[tool.mypy]
ignore_missing_imports = true
